fix some colors, fix component index
