+++ src/components/index.umd.js
@@ -1,6 +1,6 @@
 /**
  * @file index.js
  */
-import xcui from './index';
+import xcui from './index.esmodule';
 
 module.exports = xcui;
