+++ src/less/components/progress.less
@@ -6,7 +6,7 @@
 
     &-text {
         font-size:@font-size-base;
-        color:rgba(0, 0, 0, .65);
+        color: @text-color;
         display: inline-block;
         vertical-align: middle;
         margin-left: 10px;
