+++ src/less/components/tag.less
@@ -12,7 +12,7 @@
     opacity: 1;
     margin: 2px 4px 2px 0;
     cursor: pointer;
-    color: rgba(0, 0, 0, .65);
+    color: @text-color;
 
     &:hover {
         opacity: .85
